--- integration_tests/e2e/order/monitoring-conditions/add-to-list/exclusion-zone/submission.cy.ts
@@ -1,335 +1,0 @@
-import { v4 as uuidv4 } from 'uuid'
-import Page from '../../../../../pages/page'
-import EnforcementZoneAddToListPage from './ExclusionZonePage'
-import OrderSummaryPage from '../../../../../pages/order/summary'
-import { EnforcementZoneAddToListFormData } from './ExclusionZoneComponent'
-import TypesOfMonitoringNeededPage from '../../order-type-description/types-of-monitoring-needed/TypesOfMonitoringNeededPage'
-
-const mockOrderId = uuidv4()
-const apiPath = '/enforcementZone'
-const uploadApiPath = '/attachment'
-
-context('Monitoring conditions - Enforcement Zone', () => {
-  context('Submitting a valid Exclusion zone order', () => {
-    const zoneType = 'Exclusion zone'
-    const zoneTypeId = 'EXCLUSION'
-
-    beforeEach(() => {
-      cy.task('reset')
-      cy.task('stubSignIn', { name: 'john smith', roles: ['ROLE_EM_CEMO__CREATE_ORDER'] })
-
-      cy.task('stubCemoGetOrder', {
-        httpStatus: 200,
-        id: mockOrderId,
-        status: 'IN_PROGRESS',
-        order: {
-          monitoringConditions: {
-            orderType: 'IMMIGRATION',
-            orderTypeDescription: null,
-            conditionType: null,
-            acquisitiveCrime: false,
-            dapol: false,
-            curfew: false,
-            exclusionZone: false,
-            trail: false,
-            mandatoryAttendance: true,
-            alcohol: false,
-            startDate: null,
-            endDate: null,
-            sentenceType: null,
-            issp: null,
-            hdc: null,
-            prarr: null,
-            pilot: null,
-            offenceType: null,
-          },
-        },
-      })
-      cy.task('stubCemoSubmitOrder', {
-        httpStatus: 200,
-        id: mockOrderId,
-        subPath: apiPath,
-        response: {
-          orderId: mockOrderId,
-          zoneType: zoneTypeId,
-          zoneId: 1,
-          startDate: '2024-10-10',
-          endDate: '2024-10-11',
-          name: 'A test name: Lorem ipsum dolor sit amet...',
-          description: 'A test description: Lorem ipsum dolor sit amet...',
-          duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        },
-      })
-
-      cy.signIn()
-    })
-
-    it('should submit a correctly formatted Exclusion zone submission', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAndContinueButton.click()
-
-      cy.task('stubCemoVerifyRequestReceived', {
-        uri: `/orders/${mockOrderId}${apiPath}`,
-        body: {
-          zoneType: zoneTypeId,
-          zoneId: 1,
-
-          startDate: '2024-12-10T00:00:00.000Z',
-          endDate: '2024-12-11T23:59:00.000Z',
-          name: 'A test name: Lorem ipsum dolor sit amet...',
-          description: 'A test description: Lorem ipsum dolor sit amet...',
-          duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        },
-      }).should('be.true')
-    })
-
-    it('should submit a correctly without enddate when notifying organisation is HOME_OFFICE', () => {
-      cy.task('stubCemoGetOrder', {
-        httpStatus: 200,
-        id: mockOrderId,
-        status: 'IN_PROGRESS',
-        order: {
-          interestedParties: {
-            notifyingOrganisation: 'HOME_OFFICE',
-            notifyingOrganisationName: 'Home Office',
-            notifyingOrganisationEmail: 'test@homeoffice.gov.uk',
-            responsibleOfficerName: 'Test Officer',
-            responsibleOfficerPhoneNumber: '01234567890',
-            responsibleOrganisation: 'PROBATION',
-            responsibleOrganisationRegion: 'Test Region',
-            responsibleOrganisationEmail: 'test@probation.gov.uk',
-          },
-          monitoringConditions: {
-            orderType: 'IMMIGRATION',
-            orderTypeDescription: null,
-            conditionType: null,
-            acquisitiveCrime: false,
-            dapol: false,
-            curfew: false,
-            exclusionZone: false,
-            trail: false,
-            mandatoryAttendance: true,
-            alcohol: false,
-            startDate: null,
-            endDate: null,
-            sentenceType: null,
-            issp: null,
-            hdc: null,
-            prarr: null,
-            pilot: null,
-            offenceType: null,
-          },
-        },
-      })
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAndContinueButton.click()
-
-      cy.task('stubCemoVerifyRequestReceived', {
-        uri: `/orders/${mockOrderId}${apiPath}`,
-        body: {
-          zoneType: zoneTypeId,
-          zoneId: 1,
-          startDate: '2024-12-10T00:00:00.000Z',
-          name: 'A test name: Lorem ipsum dolor sit amet...',
-          description: 'A test description: Lorem ipsum dolor sit amet...',
-          duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        },
-      }).should('be.true')
-    })
-
-    it('should continue to the types of monitoring needed page', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAndContinueButton.click()
-
-      Page.verifyOnPage(TypesOfMonitoringNeededPage)
-    })
-
-    it('should return to the summary page', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAsDraftButton.click()
-
-      Page.verifyOnPage(OrderSummaryPage)
-    })
-  })
-  context('submitting an enforcement zone order with a file', () => {
-    const zoneId = 1
-    const zoneType = 'Exclusion zone'
-    const zoneTypeId = 'EXCLUSION'
-    const fileContents = 'Test image file'
-
-    beforeEach(() => {
-      cy.task('reset')
-      cy.task('stubSignIn', { name: 'john smith', roles: ['ROLE_EM_CEMO__CREATE_ORDER'] })
-
-      cy.task('stubCemoGetOrder', {
-        httpStatus: 200,
-        id: mockOrderId,
-        status: 'IN_PROGRESS',
-        order: {
-          monitoringConditions: {
-            orderType: 'IMMIGRATION',
-            orderTypeDescription: null,
-            conditionType: null,
-            acquisitiveCrime: false,
-            dapol: false,
-            curfew: false,
-            exclusionZone: false,
-            trail: false,
-            mandatoryAttendance: true,
-            alcohol: false,
-            startDate: null,
-            endDate: null,
-            sentenceType: null,
-            issp: null,
-            hdc: null,
-            prarr: null,
-            pilot: null,
-            offenceType: null,
-          },
-        },
-      })
-      cy.task('stubCemoSubmitOrder', {
-        httpStatus: 200,
-        id: mockOrderId,
-        subPath: apiPath,
-        response: {
-          orderId: mockOrderId,
-          zoneType: zoneTypeId,
-          zoneId,
-          startDate: '2024-10-10',
-          endDate: '2024-10-11',
-          name: 'A test name: Lorem ipsum dolor sit amet...',
-          description: 'A test description: Lorem ipsum dolor sit amet...',
-          duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        },
-      })
-      cy.task('stubCemoSubmitOrder', {
-        httpStatus: 200,
-        method: 'POST',
-        id: mockOrderId,
-        subPath: `${apiPath}/${zoneId}${uploadApiPath}`,
-        response: {},
-      })
-
-      cy.signIn()
-    })
-
-    it('should submit a correctly formatted file upload submission', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        uploadFile: {
-          fileName: 'test-image.jpeg',
-          rawContent: fileContents,
-        },
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAndContinueButton.click()
-
-      cy.task('stubCemoVerifyRequestReceived', {
-        uri: `/orders/${mockOrderId}${apiPath}/${zoneId}${uploadApiPath}`,
-        fileContents: [
-          {
-            name: 'file',
-            filename: 'test-image.jpeg',
-            contentType: 'image/jpeg',
-            contents: fileContents,
-          },
-        ],
-      }).should('be.true')
-    })
-
-    it('should continue to the types of monitoring needed page', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData: EnforcementZoneAddToListFormData = {
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        uploadFile: {
-          fileName: 'test-image.jpeg',
-          rawContent: fileContents,
-        },
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAndContinueButton.click()
-
-      Page.verifyOnPage(TypesOfMonitoringNeededPage, { orderId: mockOrderId, zoneId: 2 })
-    })
-
-    it('should return to the summary page', () => {
-      const page = Page.visit(EnforcementZoneAddToListPage, { orderId: mockOrderId, zoneId: 1 })
-
-      const validFormData = {
-        zoneType,
-        startDate: new Date('2024-12-10T00:00:00.000Z'),
-        endDate: new Date('2024-12-11T00:00:00.000Z'),
-        name: 'A test name: Lorem ipsum dolor sit amet...',
-        description: 'A test description: Lorem ipsum dolor sit amet...',
-        duration: 'A test duration: Lorem ipsum dolor sit amet...',
-        uploadFile: {
-          fileName: 'test-image.jpeg',
-          rawContent: fileContents,
-        },
-      }
-
-      page.form.fillInWith(validFormData)
-      page.form.saveAsDraftButton.click()
-
-      Page.verifyOnPage(OrderSummaryPage)
-    })
-  })
-})
