+++ server/constants/paths.ts
@@ -84,8 +84,8 @@ const paths = {
     INSTALLATION_APPOINTMENT: '/order/:orderId/monitoring-conditions/installation-appointment',
     TRAIL: '/order/:orderId/monitoring-conditions/trail',
     ZONE: '/order/:orderId/monitoring-conditions/zone/:zoneId',
-    ZONE_NEW_ITEM: '/order/:orderId/monitoring-conditions/add-to-list/zone',
-    ZONE_ADD_TO_LIST: '/order/:orderId/monitoring-conditions/add-to-list/zone/:zoneId',
+    ZONE_NEW_ITEM: '/order/:orderId/monitoring-conditions/add-to-list/zone/:zoneType',
+    ZONE_ADD_TO_LIST: '/order/:orderId/monitoring-conditions/add-to-list/zone/:zoneType/:zoneId',
     ATTENDANCE: '/order/:orderId/monitoring-conditions/attendance',
     ATTENDANCE_ADD_TO_LIST: '/order/:orderId/monitoring-conditions/add-to-list/attendance',
     ATTENDANCE_ITEM: '/order/:orderId/monitoring-conditions/attendance/:conditionId',
