+++ server/routes/monitoring-conditions/monitoring-type/controller.ts
@@ -49,15 +49,17 @@ export default class MonitoringTypeController {
   }
 
   getMonitoringConditionPath = (
-    condition: 'curfew' | 'exclusionZone' | 'trail' | 'mandatoryAttendance' | 'alcohol',
+    condition: 'curfew' | 'exclusionZone' | 'trail' | 'mandatoryAttendance' | 'alcohol' | 'restrictionZone',
   ): string => {
     switch (condition) {
       case 'alcohol':
         return paths.MONITORING_CONDITIONS.ALCOHOL
       case 'curfew':
         return paths.MONITORING_CONDITIONS.CURFEW_CONDITIONS
+      case 'restrictionZone':
+        return paths.MONITORING_CONDITIONS.ZONE_NEW_ITEM.replace(':zoneType', 'exclusion')
       case 'exclusionZone':
-        return paths.MONITORING_CONDITIONS.ZONE_NEW_ITEM
+        return paths.MONITORING_CONDITIONS.ZONE_NEW_ITEM.replace(':zoneType', 'restriction')
       case 'trail':
         return paths.MONITORING_CONDITIONS.TRAIL
       case 'mandatoryAttendance':
